Crude first demo - client and server daemons.
This includes both client and server daemons. When started, the server
will open the UDP socket and listen for incomming datagrams, when a GRE
packet from a new client is detected, it will create a new tunnel device
use it for incoming traffic. The client will open a UDP socket and tunnel
device and bind them to the server.

Once created, you can then issue ICMPv6 pings to the link local
addresses.