Add zero-length GRE packet for connection setup.
Without a ZLP, the setup of a tunnel interface on the server doesn't
really take place until the first packet from client to server is sent.
It makes things a lot more deterministic if connection setup occurs on
when booting the client daemon instead. Eventually we want to do this
using DTLS instead.